+++ coxeter/shape_families/doi_data_repositories.py
@@ -49,12 +49,12 @@ def _doi_shape_collection_factory(doi):
         for fn in doi_to_file[doi]:
             with open(os.path.join(_DATA_FOLDER, fn)) as f:
                 families.append(TabulatedGSDShapeFamily(json.load(f)))
-    elif doi in doi_to_family:
+    if doi in doi_to_family:
         for family_type in doi_to_family[doi]:
             families.append(family_type())
     else:
         raise KeyError(
-            "Provided DOI is not associated with any known data or " "shape families."
+            "Provided DOI is not associated with any known data or shape families."
         )
     return families